refactor: move actions to the file that corresponds to the db model they are acting upon; update related files that use these actions
