+++ src/actions/user.ts
@@ -1,7 +1,5 @@
 'use server';
 
-import { revalidatePath } from 'next/cache';
-
 import prisma from '@/lib/prisma';
 
 export async function searchUsersAction(query: string) {
@@ -38,143 +36,6 @@ export async function searchUsersAction(query: string) {
   });
 
   return userResults;
-}
-
-export async function addMembersToGroupAction(usersToAdd, groupId) {
-  const members = usersToAdd.map((user) => ({
-    user_id: user.id,
-    group_id: groupId,
-  }));
-
-  const result = await prisma.member.createMany({
-    data: members,
-    skipDuplicates: true,
-  });
-
-  revalidatePath(`/dashboard/group/${groupId}/members`);
-  return result;
-}
-
-export async function addEmployeesToGroupAction(usersToAdd, groupId) {
-  const employees = usersToAdd.map((user) => ({
-    user_id: user.id,
-    group_id: groupId,
-  }));
-
-  const result = await prisma.employee.createMany({
-    data: employees,
-    skipDuplicates: true,
-  });
-
-  revalidatePath(`/dashboard/group/${groupId}/employees`);
-  return result;
-}
-
-export async function editMemberAction(
-  { user_id, group_id, id },
-  role,
-  status,
-) {
-  // console.log('editing member', member.id, role, status);
-  try {
-    // throw new Error('test error');
-    const result = await prisma.member.update({
-      // if querying by @@unique constraint
-      // where: {
-      //   user_id_group_id: {
-      //     user_id: member.user.id,
-      //     group_id: member.group.id,
-      //   },
-      // },
-      where: { id },
-      data: {
-        role,
-        // status
-      },
-    });
-    revalidatePath(`/dashboard/group/${group_id}/members`);
-    return result;
-  } catch (err) {
-    console.error(err);
-  }
-}
-
-export async function removeMemberAction({ user_id, group_id, id }) {
-  // console.log('removing member');
-
-  const deletedCheckins = await prisma['member'].update({
-    where: { id },
-    data: {
-      checkins: {
-        deleteMany: {},
-      },
-    },
-    include: {
-      checkins: true,
-    },
-  });
-  // console.log('deleted checkins', deletedCheckins);
-
-  if (!deletedCheckins) return;
-
-  const removedMember = await prisma['member'].delete({
-    where: { id },
-    select: {
-      user_id: true,
-      user: { include: { profile: { select: { full_name: true } } } },
-    },
-  });
-  // console.log('removed member', removedMember);
-
-  revalidatePath(`/dashboard/group/${group_id}/members`);
-  return removedMember;
-}
-
-export async function editEmployeeAction(
-  { user_id, group_id },
-  role,
-  // status,
-) {
-  // console.log('editing employee action');
-  try {
-    const editedEmployee = await prisma.employee.update({
-      where: {
-        user_id_group_id: {
-          user_id,
-          group_id,
-        },
-      },
-      data: {
-        role,
-        // status
-      },
-    });
-    revalidatePath(`/dashboard/group/${group_id}/employees`);
-    return editedEmployee;
-  } catch (err) {
-    console.error(err);
-  }
-}
-
-export async function removeEmployeeAction({ user_id, group_id }) {
-  // console.log('removing employee action');
-
-  const removedEmployee = await prisma['employee'].delete({
-    where: {
-      user_id_group_id: {
-        user_id,
-        group_id,
-      },
-    },
-    select: {
-      user_id: true,
-      user: { include: { profile: { select: { full_name: true } } } },
-    },
-  });
-  console.log('removed employee', removedEmployee);
-
-  revalidatePath(`/dashboard/group/${group_id}/employee`);
-  return removedEmployee;
 }
 
 // original query based on sadmann7/skateshop repo
